requirements.txt
fastapi
uvicorn
transformers
torch
sentencepiece
python-multipart
pydantic
PyPDF2
sqlalchemy
psycopg2-binary
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.12.1
